CS5481 - Group Porject/topic2_automated_summary/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
openai==1.14.0
python-dotenv==1.0.1
newspaper3k==0.2.8
nltk==3.8.1
sentence-transformers==2.5.1
networkx==3.2.1
matplotlib==3.8.3
jinja2==3.1.3